🧹 refactor(all): refactor all components
