+++ .env.example
@@ -2,6 +2,4 @@
 # Currently, there are no environment variables needed.
 
 # Example of what future variables might look like:
-# NEXT_PUBLIC_API_KEY=your_api_key_here
-# DATABASE_URL=your_database_url_here
-# SECRET_KEY=your_secret_key_here
+TO_EMAIL="smth"
