+++ src/app/api/send-email/route.ts
@@ -33,8 +33,8 @@ export async function POST(req: NextRequest) {
         "Content-Type": "application/json",
       },
       body: JSON.stringify({
-        from: "divyanshudhruv24@gmail.com", // Use your verified sender email here
-        to: "divyanshudhruv24@gmail.com", // Use the recipient email from your environment variable
+        from: "Contact Form <onboarding@resend.dev>", // Use your verified sender email here
+        to: TO_EMAIL, // Use the recipient email from your environment variable
         subject: "New Message",
         html: `<p>Message from: ${email}</p><div>${text}</div>`,
         reply_to: email,
