+++ src/app/components/Footer.tsx
@@ -1,116 +1,18 @@
 "use client";
 
-import BlobCursor from "@/blocks/Animations/BlobCursor/BlobCursor";
 import Magnet from "@/blocks/Animations/Magnet/Magnet";
-import Waves from "@/blocks/Backgrounds/Waves/Waves";
-import Dock from "@/blocks/Components/Dock/Dock";
-import GradientText from "@/blocks/TextAnimations/GradientText/GradientText";
-import ScrollVelocity from "@/blocks/TextAnimations/ScrollVelocity/ScrollVelocity";
-import ShinyText from "@/blocks/TextAnimations/ShinyText/ShinyText";
+import { Text, Column, Row, Flex, Button } from "@once-ui-system/core";
 
-import {
-  Heading,
-  Text,
-  Button,
-  Column,
-  Badge,
-  Logo,
-  Line,
-  LetterFx,
-  Row,
-  ThemeSwitcher,
-  Flex,
-  StatusIndicator,
-  Grid,
-  Card,
-  Media,
-  Tag,
-  Input,
-  Textarea,
-  useToast,
-} from "@once-ui-system/core";
-import {
-  ArchiveIcon,
-  ArrowDown,
-  ArrowUpRight,
-  GitPullRequestIcon,
-  House,
-  MessageCircle,
-  Send,
-  SettingsIcon,
-  TagIcon,
-} from "lucide-react";
+import { Instrument_Serif, Poppins, Inter, Geist_Mono } from "next/font/google";
 
-import {
-  Instrument_Serif,
-  Poppins,
-  Inter,
-  Advent_Pro,
-  Archivo_Narrow,
-  Roboto_Serif,
-  Noto_Serif,
-  Source_Serif_4,
-  PT_Serif,
-  Geist_Mono,
-} from "next/font/google";
-import { BiArchive, BiHome } from "react-icons/bi";
-import { GiSettingsKnobs } from "react-icons/gi";
-import { MdAccountBalanceWallet } from "react-icons/md";
-import { useEffect, useRef, useState } from "react";
-import FlowingMenu from "@/blocks/Components/FlowingMenu/FlowingMenu";
-import BounceCards from "@/blocks/Components/BounceCards/BounceCards";
-import LightRays from "@/blocks/Backgrounds/LightRays/LightRays";
-import { IoArrowDownSharp } from "react-icons/io5";
-import Lenis from "lenis";
-import gsap from "gsap";
-import { ScrollTrigger } from "gsap/ScrollTrigger";
-import ClickSpark from "@/blocks/Animations/ClickSpark/ClickSpark";
-import CircularText from "@/blocks/TextAnimations/CircularText/CircularText";
-import Threads from "@/blocks/Backgrounds/Threads/Threads";
-import Orb from "@/blocks/Backgrounds/Orb/Orb";
+import { useEffect, useState } from "react";
+
 import React from "react";
-const instrument_serif = Instrument_Serif({
-  weight: ["400"],
-  subsets: ["latin"],
-});
-const poppins = Poppins({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
-const inter = Inter({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
+
 
 const geist_mono = Geist_Mono({
   weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
   subsets: ["latin"],
-});
-
-const advent_pro = Advent_Pro({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
-const archivo_narrow = Archivo_Narrow({
-  weight: ["400", "700", "600", "500"],
-  subsets: ["latin"],
-});
-const roboto_serif = Roboto_Serif({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
-const noto_serif = Noto_Serif({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-
-  subsets: ["latin"],
-});
-const source_serif_4 = Source_Serif_4({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200"],
-  subsets: ["latin"],
-});
-const pt_serif = PT_Serif({
-  subsets: ["latin"],
-  weight: ["400", "700"],
 });
 
 export default function Footer() {
@@ -130,7 +32,7 @@ export default function Footer() {
         gap="80"
         paddingY="16"
       >
-        {/* <div
+         <div
             style={{
               backgroundColor: "transparent",
               height: "50px",
@@ -165,7 +67,6 @@ export default function Footer() {
           <Text
             variant="code-default-s"
             className={geist_mono.className}
-            marginBottom="64"
           >
             {" "}
             <Row
@@ -242,12 +143,12 @@ export default function Footer() {
                 transform: "translate(-50%, -50%)",
               }}
             >
-               <Orb
+               {/* <Orb
                 hoverIntensity={0.6}
                 rotateOnHover={true}
                 hue={0}
                 forceHoverState={false}
-              /> 
+              />  */}
             </div>
             <Text
               className={geist_mono.className}
@@ -261,7 +162,7 @@ export default function Footer() {
               }}
             >
               A portfolio by
-              <br />☻ Divyanshu Dhruv ☻
+              <br />me
             </Text>
             <Magnet magnetStrength={10}>
               <Button
@@ -294,7 +195,7 @@ export default function Footer() {
                 &nbsp;Github
               </Button>
             </Magnet>
-          </Column> */}{" "}
+          </Column> 
         <Column fillWidth>
           <Row
             fillWidth
