+++ src/app/components/Projects.tsx
@@ -1,73 +1,32 @@
 "use client";
 
-import BlobCursor from "@/blocks/Animations/BlobCursor/BlobCursor";
 import Magnet from "@/blocks/Animations/Magnet/Magnet";
-import Waves from "@/blocks/Backgrounds/Waves/Waves";
-import Dock from "@/blocks/Components/Dock/Dock";
-import GradientText from "@/blocks/TextAnimations/GradientText/GradientText";
+
 import ScrollVelocity from "@/blocks/TextAnimations/ScrollVelocity/ScrollVelocity";
 import ShinyText from "@/blocks/TextAnimations/ShinyText/ShinyText";
 
 import {
-  Heading,
   Text,
   Button,
   Column,
-  Badge,
-  Logo,
-  Line,
-  LetterFx,
   Row,
-  ThemeSwitcher,
   Flex,
   StatusIndicator,
   Grid,
   Card,
   Media,
   Tag,
-  Input,
-  Textarea,
-  useToast,
 } from "@once-ui-system/core";
-import {
-  ArchiveIcon,
-  ArrowDown,
-  ArrowUpRight,
-  GitPullRequestIcon,
-  House,
-  MessageCircle,
-  Send,
-  SettingsIcon,
-  TagIcon,
-} from "lucide-react";
+import { ArrowUpRight } from "lucide-react";
 
-import {
-  Instrument_Serif,
-  Poppins,
-  Inter,
-  Advent_Pro,
-  Archivo_Narrow,
-  Roboto_Serif,
-  Noto_Serif,
-  Source_Serif_4,
-  PT_Serif,
-  Geist_Mono,
-} from "next/font/google";
-import { BiArchive, BiHome } from "react-icons/bi";
-import { GiSettingsKnobs } from "react-icons/gi";
-import { MdAccountBalanceWallet } from "react-icons/md";
-import { useEffect, useRef, useState } from "react";
-import FlowingMenu from "@/blocks/Components/FlowingMenu/FlowingMenu";
-import BounceCards from "@/blocks/Components/BounceCards/BounceCards";
-import LightRays from "@/blocks/Backgrounds/LightRays/LightRays";
+import { Instrument_Serif, Poppins, Inter, Geist_Mono } from "next/font/google";
+import { useEffect, useRef } from "react";
+
 import { IoArrowDownSharp } from "react-icons/io5";
-import Lenis from "lenis";
 import gsap from "gsap";
+
 import { ScrollTrigger } from "gsap/ScrollTrigger";
-import ClickSpark from "@/blocks/Animations/ClickSpark/ClickSpark";
-import CircularText from "@/blocks/TextAnimations/CircularText/CircularText";
-import Threads from "@/blocks/Backgrounds/Threads/Threads";
-import Orb from "@/blocks/Backgrounds/Orb/Orb";
+
 import React from "react";
 const instrument_serif = Instrument_Serif({
   weight: ["400"],
@@ -87,32 +46,6 @@ const geist_mono = Geist_Mono({
   subsets: ["latin"],
 });
 
-const advent_pro = Advent_Pro({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
-const archivo_narrow = Archivo_Narrow({
-  weight: ["400", "700", "600", "500"],
-  subsets: ["latin"],
-});
-const roboto_serif = Roboto_Serif({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
-const noto_serif = Noto_Serif({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-
-  subsets: ["latin"],
-});
-const source_serif_4 = Source_Serif_4({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200"],
-  subsets: ["latin"],
-});
-const pt_serif = PT_Serif({
-  subsets: ["latin"],
-  weight: ["400", "700"],
-});
-
 export default function Projects() {
   return (
     <>
@@ -130,7 +63,6 @@ export default function Projects() {
         paddingX="m"
         gap="128"
       >
-        <LocoScrollImg />
         <Text className={geist_mono.className} style={{ fontWeight: "200" }}>
           {" "}
           <ScrollVelocity
@@ -273,92 +205,7 @@ export default function Projects() {
   );
 }
 
-function LocoScrollImg() {
-  gsap.registerPlugin(ScrollTrigger);
 
-  // 6 images, 3 per side, fixed order
-  const images = [
-    { src: "/donut.svg", alt: "Donut" },
-    { src: "/pyramid.svg", alt: "Pyramid" },
-    { src: "/pill.svg", alt: "Pill" },
-    { src: "/sphere.svg", alt: "Sphere" },
-    { src: "/cube.svg", alt: "Cube" },
-    { src: "/cone.svg", alt: "Cone" },
-  ];
-  const count = 3;
-  const sides = ["left", "right"];
-  // Fixed speed factors for each image (can be customized)
-  const speedFactorsBySide = [
-    [1.5, 2.0, 2.5],
-    [1.8, 2.2, 2.7],
-  ];
-
-  // Refs for all images
-  const imgRefs = useRef<(HTMLImageElement | null)[][]>(
-    Array.from({ length: sides.length }, () => Array(count).fill(null))
-  );
-
-  useEffect(() => {
-    imgRefs.current.forEach((sideRefs, sideIdx) => {
-      sideRefs.forEach((img, idx) => {
-        if (!img) return;
-        const speed = speedFactorsBySide[sideIdx][idx];
-        gsap.to(img, {
-          y: () => `-${speed * 200}px`,
-          ease: "none",
-          scrollTrigger: {
-            trigger: img,
-            start: "top bottom",
-            end: "bottom top",
-            scrub: true,
-          },
-        });
-      });
-    });
-    return () => {
-      ScrollTrigger.getAll().forEach((st) => st.kill());
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  // Assign first 3 images to left, next 3 to right, fixed order
-  return (
-    <>
-      {sides.map((side, sideIdx) => {
-        const imgs = images.slice(sideIdx * count, sideIdx * count + count);
-        const speedFactors = speedFactorsBySide[sideIdx];
-        return imgs.map((img, idx) => {
-          const baseGap = 30; // percent
-          const offset = side === "left" ? 10 : 20;
-          const top = `${offset + idx * baseGap}%`;
-          const size = 125 + speedFactors[idx] * 85; // px
-          return (
-            <img
-              key={side + "-" + img.src + "-" + idx}
-              ref={(el) => {
-                imgRefs.current[sideIdx][idx] = el;
-              }}
-              src={img.src}
-              alt={img.alt}
-              style={{
-                position: "absolute",
-                [side]: 0,
-                top,
-                width: size,
-                height: size,
-                zIndex: 1,
-                pointerEvents: "none",
-                opacity: 1,
-                userSelect: "none",
-              }}
-              draggable={true}
-            />
-          );
-        });
-      })}
-    </>
-  );
-}
 
 type ProjectsProps = {
   title: string;
