+++ src/app/components/Skills.tsx
@@ -1,117 +1,31 @@
 "use client";
 
-import BlobCursor from "@/blocks/Animations/BlobCursor/BlobCursor";
 import Magnet from "@/blocks/Animations/Magnet/Magnet";
-import Waves from "@/blocks/Backgrounds/Waves/Waves";
-import Dock from "@/blocks/Components/Dock/Dock";
-import GradientText from "@/blocks/TextAnimations/GradientText/GradientText";
-import ScrollVelocity from "@/blocks/TextAnimations/ScrollVelocity/ScrollVelocity";
+
 import ShinyText from "@/blocks/TextAnimations/ShinyText/ShinyText";
 
-import {
-  Heading,
-  Text,
-  Button,
-  Column,
-  Badge,
-  Logo,
-  Line,
-  LetterFx,
-  Row,
-  ThemeSwitcher,
-  Flex,
-  StatusIndicator,
-  Grid,
-  Card,
-  Media,
-  Tag,
-  Input,
-  Textarea,
-  useToast,
-} from "@once-ui-system/core";
-import {
-  ArchiveIcon,
-  ArrowDown,
-  ArrowUpRight,
-  GitPullRequestIcon,
-  House,
-  MessageCircle,
-  Send,
-  SettingsIcon,
-  TagIcon,
-} from "lucide-react";
+import { Text, Button, Column, Row, Flex } from "@once-ui-system/core";
+import { ArrowUpRight } from "lucide-react";
 
-import {
-  Instrument_Serif,
-  Poppins,
-  Inter,
-  Advent_Pro,
-  Archivo_Narrow,
-  Roboto_Serif,
-  Noto_Serif,
-  Source_Serif_4,
-  PT_Serif,
-  Geist_Mono,
-} from "next/font/google";
-import { BiArchive, BiHome } from "react-icons/bi";
-import { GiSettingsKnobs } from "react-icons/gi";
-import { MdAccountBalanceWallet } from "react-icons/md";
-import { useEffect, useRef, useState } from "react";
-import FlowingMenu from "@/blocks/Components/FlowingMenu/FlowingMenu";
+import { Instrument_Serif, Inter } from "next/font/google";
+
 import BounceCards from "@/blocks/Components/BounceCards/BounceCards";
-import LightRays from "@/blocks/Backgrounds/LightRays/LightRays";
 import { IoArrowDownSharp } from "react-icons/io5";
-import Lenis from "lenis";
-import gsap from "gsap";
-import { ScrollTrigger } from "gsap/ScrollTrigger";
-import ClickSpark from "@/blocks/Animations/ClickSpark/ClickSpark";
-import CircularText from "@/blocks/TextAnimations/CircularText/CircularText";
-import Threads from "@/blocks/Backgrounds/Threads/Threads";
-import Orb from "@/blocks/Backgrounds/Orb/Orb";
-import React from "react";
+
+import React, { useEffect, useRef } from "react";
 const instrument_serif = Instrument_Serif({
   weight: ["400"],
   subsets: ["latin"],
 });
-const poppins = Poppins({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
+
 const inter = Inter({
   weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
   subsets: ["latin"],
 });
 
-const geist_mono = Geist_Mono({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
+import gsap from "gsap";
+import { ScrollTrigger } from "gsap/ScrollTrigger";
 
-const advent_pro = Advent_Pro({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
-const archivo_narrow = Archivo_Narrow({
-  weight: ["400", "700", "600", "500"],
-  subsets: ["latin"],
-});
-const roboto_serif = Roboto_Serif({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-  subsets: ["latin"],
-});
-const noto_serif = Noto_Serif({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200", "100"],
-
-  subsets: ["latin"],
-});
-const source_serif_4 = Source_Serif_4({
-  weight: ["400", "700", "800", "900", "600", "500", "300", "200"],
-  subsets: ["latin"],
-});
-const pt_serif = PT_Serif({
-  subsets: ["latin"],
-  weight: ["400", "700"],
-});
 const images_row1 = [
   "https://skillicons.dev/icons?i=html",
   "https://skillicons.dev/icons?i=css",
@@ -151,18 +65,7 @@ const images_row3 = [
   "https://skillicons.dev/icons?i=vscode",
   "https://skillicons.dev/icons?i=bash",
 ];
-const education_row1 = [
-  "/spa.png",
 
-  "https://i.pinimg.com/736x/1d/5b/65/1d5b651263a4c831fa6ef4e28bef3eaf.jpg",
-  "https://tse4.mm.bing.net/th/id/OIP.6ApXHS-og78F4IEE0VcLrQHaHa?r=0&rs=1&pid=ImgDetMain&o=7&rm=3",
-];
-
-const education_row1_styles = [
-  "rotate(5deg) translate(-100px)",
-  "rotate(0deg) translate(0px)",
-  "rotate(-3deg) translate(100px)",
-];
 const images_row4 = [
   "https://skillicons.dev/icons?i=codepen",
   "https://skillicons.dev/icons?i=arduino",
@@ -213,12 +116,6 @@ const getTransformStyles = (images: string[]) =>
 const transformStyles_row1 = getTransformStyles(images_row1);
 const transformStyles_row2 = getTransformStyles(images_row2);
 const transformStyles_row3 = getTransformStyles(images_row3);
-const transformStyles_row4 = getTransformStyles(images_row4);
-const transformStyles_row5 = getTransformStyles(images_row5);
-
-const lenis = new Lenis({
-  autoRaf: true,
-});
 
 export default function Skills() {
   return (
@@ -234,8 +131,8 @@ export default function Skills() {
         padding="m"
         paddingX="m"
         gap="128"
+        marginTop="xl"
       >
-        {" "}
         <Column
           fillWidth
           horizontal="center"
